requirement.txt
fastapi
uvicorn
ultralytics
facenet-pytorch
torch
pillow
requests
numpy